Fix Chef/WhyRunSupportedTrue to find the correct method
:facepalm: I got the wrong method here.
