+++ config/cookstyle.yml
@@ -270,7 +270,7 @@ Chef/LegacyBerksfileSource:
     - '**/Berksfile'
 
 Chef/WhyRunSupportedTrue:
-  Description: why_run_supported? no longer needs to be set to true as it is the default in Chef 13+
+  Description: whyrun_supported? no longer needs to be set to true as it is the default in Chef 13+
   Enabled: true
   VersionAdded: '5.1.0'
   Include:
