+++ test.rb
@@ -1,0 +1,9 @@
+# node.save
+
+use_inline_resources
+
+# use_inline_resources if defined?(use_inline_resources)
+
+use_inline_resources if defined?(use_inline_resources2)
+
+# (if (defined? (send nil :use_inline_resources)) (send nil :use_inline_resources) nil)
